Use Polymer.Gestures helper to listen for event called tap
In Polymer 1.x we would use an on-tap="functionCallback" declarative
listener in a component template to listen for click and tap
(mobile friendly click events for old browsers). Because of changes
in dropdown we need to add event listeners with JavaScript now.

This tries to catch all kinds of click events correctly using the
Polymer.Gestures library, which works in Polymer 1 or 2.